added react-firebaseui and refacted login with redux
