+++ package.json
@@ -9,6 +9,7 @@
     "font-awesome": "^4.7.0",
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
+    "react-firebaseui": "^3.0.7",
     "react-redux": "^5.0.7",
     "react-router-dom": "^4.3.1",
     "react-scripts": "1.1.5",
