+++ src/components/NavBar/actions.js
@@ -1,6 +1,5 @@
 import { LOGOUT } from '../Auth/reducers';
 import { auth } from '../../services/firebase';
-// export const logout = () => ({ type: LOGOUT });
 
 export const logout = () => dispatch => {
   return auth.doSignOut(() => dispatch({ type: LOGOUT }))
