+++ src/index.js
@@ -1,27 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { compose, createStore, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
-import thunk from 'redux-thunk';
-import { createLogger } from 'redux-logger';
-import rootReducer from './rootReducer';
 import App from './components/App/App';
 import registerServiceWorker from './registerServiceWorker';
+import store from './store';
 import './index.css';
 
 registerServiceWorker();
-
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
-const middleware = [thunk];
-
-if (process.env.NODE_ENV !== 'production') {
-  middleware.push(createLogger())
-}
-
-const store = createStore(
-  rootReducer,
-  composeEnhancers(applyMiddleware(...middleware))
-);
 
 ReactDOM.render(
   <Provider store={store}>
